nvlist pack/unpack tester
Add an nvpair_packed test command which verifies that an nvlist packed
in the XDR format can be unpacked to an identical list and (when -r
<refdir> is specified) the packed output is identical to a reference
packed output stored in <refdir>/<case>.ref.  Reference output can be
generated by adding the -R option.

The core set of tests is defined in `struct nvcase test_cases[]`, a
statically initialized array describing nvpairs to generate.  A series
of macros are used to keep the verbosity under control.  The main
exception is embedded nvlists which are initialized in init_nvlists().
The current set is in data_type_t order and all types are coverd, but
some edge cases may be unexplored, especially since only a small number
of multiple-element nvlists are created.

Signed-off-by: Brooks Davis <brooks.davis@sri.com>