nvpair: Always use XDR encoding
Recievers are agnostic to stream format and just call nvpair_unpack or
nvpair_xunpack which decode what ever stream type is received.  It is
easier to make a system with different unpacked sizes work with XDR than
with native where streams are unpacked in place.

This change provides forward compatability with systems were the size of
pointers exceeds 64-bits.  For example, a CheriBSD pure-capability
kernel targeting Morello should be able to work with a FreeBSD aarch64
zfs or zpool command build with this change applied.

Signed-off-by: Brooks Davis <brooks.davis@sri.com>