CHERI nvpair: support stable XDR serialization
The XDR format is agnostic to the shape of nvlist_t and nvpair_t
except that is stores the decoded size of nvpairs (which can include
embedded pointers and nvlist_ts).  In a previous commit we've taught
CHERI consumers how to read XDR streams with 64-bit pointer sizes.

This commit teaches the XDR code to compute the size the object
would be.  I've used a subtractive approach where I remove excess
padding and pointer-related sizes from the claimed size.  An
alternative approach would be to alwasy compute the nominal size
(perhaps asserting it on non-CHERI platforms).