nvlist: actually match xdrproc_t on FreeBSD
The use of a variadic function signature worked in practice on
existing platforms because the first serveral arguments are passed
in registers.  On CHERI this can break as we pass variadic arguments
in a seperate array.

NB: this should be dropped once CheriBSD incorporates a switch to a
two argument xdrproc_t.